+++ lib/Api/PaymentTransactionsApi.php
@@ -780,6 +780,114 @@ class PaymentTransactionsApi
     }
 
     /**
+     * Operation increaseAmount
+     *
+     * POST Payment/Transactions/{paymentTransactionId}/IncreaseAmount
+     *
+     * @param string $payment_transaction_id Payment ID (PCI_...) (required)
+     * @param \Secuconnect\Client\Model\PaymentTransactionsIncreaseAmountDTO $body increase amount input params 
+     * @throws ApiException on non-2xx response
+     * @return \Secuconnect\Client\Model\PaymentTransactionsProductModel
+     */
+    public function increaseAmount($payment_transaction_id, $body)
+    {
+        list($response) = $this->increaseAmountWithHttpInfo($payment_transaction_id, $body);
+        return $response;
+    }
+
+    /**
+     * Operation increaseAmountWithHttpInfo
+     *
+     * POST Payment/Transactions/{paymentTransactionId}/IncreaseAmount
+     *
+     * @param string $payment_transaction_id Payment ID (PCI_...) (required)
+     * @param \Secuconnect\Client\Model\PaymentTransactionsIncreaseAmountDTO $body increase amount input params 
+     * @throws ApiException on non-2xx response
+     * @return array of \Secuconnect\Client\Model\PaymentTransactionsProductModel, HTTP status code, HTTP response headers (array of strings)
+     */
+    public function increaseAmountWithHttpInfo($payment_transaction_id, $body)
+    {
+        // verify the required parameter 'payment_transaction_id' is set
+        if ($payment_transaction_id === null || (is_array($payment_transaction_id) && count($payment_transaction_id) === 0)) {
+            throw new \InvalidArgumentException(
+                'Missing the required parameter $payment_transaction_id when calling increaseAmount'
+            );
+        }
+        // parse inputs
+        $resourcePath = "/Payment/Transactions/{paymentTransactionId}/IncreaseAmount";
+        $httpBody = '';
+        $queryParams = [];
+        $headerParams = [];
+        $formParams = [];
+        $_header_accept = $this->apiClient->selectHeaderAccept(['application/json']);
+        if (!is_null($_header_accept)) {
+            $headerParams['Accept'] = $_header_accept;
+        }
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(['application/json']);
+
+        // path params
+        if ($payment_transaction_id !== null) {
+            $resourcePath = str_replace(
+                "{" . "paymentTransactionId" . "}",
+                $this->apiClient->getSerializer()->toPathValue($payment_transaction_id),
+                $resourcePath
+            );
+        }
+        // body params
+        $_tempBody = null;
+        if (isset($body)) {
+            $_tempBody = $body;
+        }
+
+        // for model (json/xml)
+        if (isset($_tempBody)) {
+            $httpBody = $_tempBody; // $_tempBody is the method argument, if present
+        } elseif (count($formParams) > 0) {
+            $httpBody = $formParams; // for HTTP post (form)
+        }
+        for ($retries = 0; ; $retries++) {
+
+            // this endpoint requires OAuth (access token)
+            if (strlen($this->apiClient->getConfig()->getAccessToken()) !== 0) {
+                $headerParams['Authorization'] = 'Bearer ' . $this->apiClient->getConfig()->getAccessToken();
+            }
+
+            // make the API Call
+            try {
+                list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
+                    $resourcePath,
+                    'POST',
+                    $queryParams,
+                    $httpBody,
+                    $headerParams,
+                    '\Secuconnect\Client\Model\PaymentTransactionsProductModel',
+                    '/Payment/Transactions/{paymentTransactionId}/IncreaseAmount'
+                );
+
+                return [$this->apiClient->getSerializer()->deserialize($response, '\Secuconnect\Client\Model\PaymentTransactionsProductModel', $httpHeader), $statusCode, $httpHeader];
+            } catch (ApiException $e) {
+                switch ($e->getCode()) {
+                    case 200:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\PaymentTransactionsProductModel', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                    case 401:
+                        if ($retries < 1) {
+                            Authenticator::reauthenticate();
+                            continue 2;
+                        }
+                    default:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\ProductExceptionPayload', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                }
+
+                throw $e;
+            }
+        }
+    }
+
+    /**
      * Operation revokeAccrual
      *
      * POST Payment/Transactions/{paymentTransactionId}/revokeAccrual