+++ CHANGELOG.md
@@ -19,6 +19,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 
+## [2.12.0] - 2022-02-28
+[2.12.0]:https://github.com/secuconnect/secuconnect-php-sdk/compare/2.11.0...2.12.0
+
+### Added
+- `Api.PaymentTransactionsApi.increaseAmount()`
+- `Model.PaymentTransactionsIncreaseAmountDTO`
+- `Model.SmartTransactionsProductModel` element `subscription`
+- `Model.SmartTransactionsBasketProduct` element `plan_id`
+- `Model.SmartTransactionsBasketProduct` element `start_at`
+- `Model.SmartTransactionsApplicationContextReturnUrls` element `url_push`
+
+### Fixed
+
+- docblock comment by @krsriq in [PR #19](https://github.com/secuconnect/secuconnect-php-sdk/pull/19)
+
+
 ## [2.11.0] - 2022-01-25
 [2.11.0]:https://github.com/secuconnect/secuconnect-php-sdk/compare/2.10.0...2.11.0